x2goserver/Makefile: build up sed argument line as an array and use one substitution run.
Multiple runs with the template file as input will lead to only the last
replacement to stick in the resulting generated script, which is not
what we want to have.

Correctly doing that with multiple runs would require the usage of
temporary files, which is needlessly complicated in this context.